0.9.6 code update and release
Updated:
Weather chatbot code to check if weather or season changes are actually enabled.
Removed Force Blizzard as it was redundent.
Shortened the chatbot response code a little.
Added some new const for the chatbot to refer back to.
Added a new option in the config to enable or disable the chatbot.