Nitpick push
Changed how saved weather times are loaded.  Its still a random number but now the upper bounds is the saved weather time left, lower bounds of zero to simulate the passage of time.  Previously it could allow a weather to last indefinitely.

Added in a seasonNameMap, so that localization is not in all caps, was driving me nuts.

Fixed the formatting of 3.11 weathertypes.ts file because it was driving me crazy.